Make the rustls dependency optional
This commit puts the load_native_certs function
behind a cargo feature that's turned on by default.

Fixes #2
