+++ Cargo.toml
@@ -11,13 +11,16 @@ repository = "https://github.com/ctz/rustls-native-certs"
 categories = ["network-programming", "cryptography"]
 
 [dependencies]
-rustls = "0.18.0"
+rustls = { version = "0.18.0", optional = true }
 
 [dev-dependencies]
 webpki = "0.21"
 webpki-roots = "0"
 ring = "0.16.5"
 untrusted = "0.7.0"
+
+[features]
+default = ["rustls"]
 
 [target.'cfg(windows)'.dependencies]
 schannel = "0.1.15"
