+++ admin/pipelines/cargo-steps.yml
@@ -1,6 +1,8 @@
 steps:
   - script: cargo build
     displayName: "cargo build (debug; default features)"
+  - script: cargo build --no-default-features
+    displayName: "cargo build (debug; no default features)"
   - script: cargo test
     displayName: "cargo test (debug; default features)"
     env: { "RUST_BACKTRACE": "1" }
